fix: preserve unknown metadata fields, extract compaction module, harden query topK
- compactVectorMetadata now passes metadata through byte-for-byte when within
  budget, reducing description/tags first when oversized and only falling back
  to the six-field schema when necessary
- move compaction helpers to cloudflare-worker/src/metadata.ts so the app test
  suite no longer drags workers-types globals into tsconfig.app.json
- add compactUpsertVectors boundary helper with handler-level regression tests
- clamp and validate topK in /query (Vectorize V2 caps at 50 with metadata) and
  cap findSimilarRepositories' topK+1 margin at the same limit
